inputcapture: Factor out call_dispose() from call_free()
This makes call_dispose() idempotent, so that it can be called whenever a
Call has become irrelevant, but without freeing the memory used for the
Call structure itself, which would leave it as a dangling pointer if it
is still in use as the user-data of an async call.

This will be used in a subsequent commit to keep the Call alive via a
circular reference between the GTask and Call while we are still waiting
for a result, and then break the circular reference with call_dispose()
when the Call has reached a resolution, either success or failure.

Signed-off-by: Simon McVittie <smcv@debian.org>