inputcapture: Reference-count Call instances
To avoid a use-after-free (#169), we need to make sure that there is
at least one reference held to the memory used to store the Call struct
for the duration of each D-Bus method call, so consistently take a new
ref to the user-data of g_dbus_connection_call(), and release it in
call_returned().

However, we also need to keep the Call alive for as long as it is
listening to any D-Bus signals, without introducing a long-term
circular reference that would make it be leaked indefinitely.

To achieve this, create a temporary circular reference between the Call
and the GTask, but break it via call_dispose() every time we reach a
resolution to the task, whether that's success or a failure.

Resolves: https://github.com/flatpak/libportal/issues/169
Resolves: https://github.com/flatpak/libportal/issues/190
Signed-off-by: Simon McVittie <smcv@debian.org>